Fix -Wsign-compare warning after https://codereview.chromium.org/1557133002
BUG=570908,82385
TBR=rsleevi@chromium.org

Review URL: https://codereview.chromium.org/1683673002 .

Cr-Commit-Position: refs/heads/master@{#374359}

Former-commit-id: 92600e9c8db99a4bf85cb3489741731bbd7f892e

+++ net/cert/cert_verify_proc_win.cc
@@ -804,7 +804,7 @@ CertDllVerifyRevocationWithCRLSet(DWORD encoding_type,
       }
       if (issuer_cert)
         break;
-      if (GetLastError() == CRYPT_E_SELF_SIGNED) {
+      if (GetLastError() == static_cast<DWORD>(CRYPT_E_SELF_SIGNED)) {
         issuer_cert.reset(CertDuplicateCertificateContext(subject_cert));
         break;
       }
